fix: high impact batch — dead code, dedup, cleanup (#14,#17,#19,#101)
- PlayerBar: remove dead FullscreenPlayer import and fsOpen state (#19)
- Extract shuffleArray to @/lib/utils, remove 5 local copies (#17,#101)
- Extract formatTotalDuration to @/lib/utils, remove 4 local copies (#101)
- Close #14: usePlayer() no longer used outside PlayerContext (resolved by PR #48)
